Adding a visit counter to the index page.
Summary: A class variable belongs to the class, flask initializes each view on the visit so the class has to own the counter. This is also true for the list, but since lists are a bit special (it is actually a referece) this points to the same data for all instances. (i.e it is a hack) using a class variable is the correct way to solve the problem.

Test Plan: mvn test

Reviewers: fronko

Reviewed By: fronko

Differential Revision: http://phabricator.molflow.com/D30