Fixed the indentation error in the pom file, and simplified the file transfer function in test_datamodel.py
Summary:
Regarding the file transfer function in the test_datamodel file there are several options. The previous version used the io.BytesIO
function to store the file data as a stream in the RAM. The open('file','rb') does the exact same thing as far as I can tell. Now the
rather weird dictionary that is constructed to send the file is designed in such a way that the underlying Werkzueg code can recognize it
as a filestorage object (http://werkzeug.pocoo.org/docs/0.10/datastructures/#werkzeug.datastructures.FileStorage), i.e. it must have a
filename appended at the end of the data stream. This is one of two options to send the data so that it ends up in the request.files object.
The second way is to send a string (http://werkzeug.pocoo.org/docs/0.10/test/#werkzeug.test.EnvironBuilder), but I have not tried this, as
one will always have to specify the content_type. Now we can send data in a more crude way by opening and reading a binary file to fdata and
sending it using self.app.post('/upload',data=fdata), which will cause it to end up in requests.data (which is where data ends up if flask
doesn't recognize the type) where it can be save directly. But I have not been able to do the same with the curl program that I am using
for the matlab upload function (curl seems to be a bit to smart for my taste), so we will not use this option right now.

Test Plan: None

Reviewers: joakim

Subscribers: fronko

Differential Revision: http://phabricator.molflow.com/D35