fixed bug in read_mls and read_smiles. Earlier they could only read level2 ozone files. Now they read other species files as well.
Updated newdonalettyEcmwfNC and views so the
the geolocation is included in the returned data structure.

Updated level1b_scandata_exporter_v2.py
Added two new items in the returned Frequency dictionary:
AppliedDopplerCorr and ChannelsID.
Removed SkyFreq, RestFreq, and LOFreq from returned data structure.
Decreased the number of digits in Spectra and ZeroLagVar.
Changed the length of Apodization item so it is now correct.