Implement level2 database
Summary:
The level2 data are stored in two mongodb collections: L2 and L2i.

It is not possible to post the same scan id/freq mode when it already exist in the database,
delete must be called first. This is a manual step for now.

More indexes will be needed and get endpoints will need to be implemented when the use cases
have been defined.

Docker compose and the test use the official mongo docker image.

Test Plan: arc unit

Reviewers: joakim, skymandr

Reviewed By: skymandr

Differential Revision: http://phabricator.molflow.com/D67