Always access h5py and netcdf4 files serially
Summary:
A global threading lock is used to achieve this.

There is one place, in newdonalettyERANC, where several netcdf4 files are opened at the same
time in the same thread. It has worked before though, so leaving it as it is.

Also Remove unused ecmwf modules.

Test Plan: arc unit, but note that the tests do not cover everything in this diff.

Reviewers: joakim, bengt.rydberg

Reviewed By: joakim, bengt.rydberg

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D176