Fixed so only selecting when expanding a row on overview
Summary:
Before, collapsing a row caused it to be selected which was
odd if you had more than one row expanded.

Now only expanding causes selection.

Added css for row shown in scan information so that there's a
highligh on the row that is selected.

Test Plan: Manual clicking around.

Reviewers: #odin, joakim

Reviewed By: #odin, joakim

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D433