Workaround for matplotlib caching issues
Summary:
Matplotlib > 2.x has a caching functionnality which is not thread safe and is
causing intermitent build failure when running with gunicorn.

Setting the `TRAVIS` environment variable is a workaround they added for their
own CI job.  If I understand correctly, it bypass the caching mechanism which
should remove the problem at the cost of a slight increase in startup time (not
much as using containers means that the cache don't persist between restart
anyway).

Test Plan: Ran the test suite 10 times. 40% failures before, 0 after.

Reviewers: skymandr

Reviewed By: skymandr

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D439