Running gevent with gunicorn.
Summary: After googling around I've seen others with the same 'critical timouts' in their gunicorn logs - some say this problem dissapears by using gevent. Note that gevent can run with threading but this is not enabled (we are running 4 workers).

Test Plan: Look at the code. I think this problem will only appear on a system with a high load.

Reviewers: #odin, fredrik

Reviewed By: #odin, fredrik

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D473