"Skip" the problematic test in odin API
Summary:
One of the selenium tests is failing because of a problem beween the current
versions of chromium and chromedriver in ubuntu.
I openned a ticket on launchpad but in the meantime, and since there isn't a
lot of activity in the odin repository, I suggest that it is enough to
mark the test as failing.
Using `strict=True`, whenever chromedriver is updated and the test start
passing again, it should make the job fail, prompting us to remove it.

Test Plan: n/a

Reviewers: #odin, skymandr

Reviewed By: #odin, skymandr

Subscribers: skymandr, jenkins

Differential Revision: https://phabricator.molflow.com/D548