Hide docker-compose logs
Summary:
The docker-compose outputs completely drowns pytest's with a lot of mostly useless stuff.
This redirects it to a temporary file.
The file path is printed on stderr which is shown in case of test failure so
the logs are available if needed.

Test Plan: Run tests without getting crazy

Reviewers: #odin, bengt.rydberg

Reviewed By: bengt.rydberg

Subscribers: bengt.rydberg, jenkins

Differential Revision: https://phabricator.molflow.com/D681