bug-fix of get level2 products endpoint, added some new test of endpoint.
Summary:
bug-fix of get level2 products endpoint
freqmode was previously described as a parameter to the
/rest_api/{version}/level2/development/{project}/products/
endpoint (return products for a given project)
in the swagger documentation, but freqmode was not
used by the endpoint.
Now freqmode is removed as parameter to the endpoint.

And we had another endpoint:
/rest_api/{version}/level2/development/{project}/{freqmode}/products/
that was supposed to return products for a given project and freqmode
but the result was always empty.
Now it should return the products.

Test Plan: ./run_all_tests.sh

Reviewers: skymandr, joakim

Reviewed By: skymandr

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D677