Add endpoint for project annotations
Summary:
Add a new endpoint `/rest_api/v5/level2/<project>/annotations`. This endpoint supports two methods:
- `GET` returns a list of the annotations for a given project
- `POST` allow a user with the secret key to add an annotation

Annotations consist of a `Text` and an optional `FreqMode`. When an annotation is added, the current time is included in the record as `Entry`.

The `GET` method for the new endpoint is added to swagger.

Resolve T781.

Test Plan:
- Add unit tests for the database
- Add system tests for the API

Reviewers: #odin, bengt.rydberg

Reviewed By: bengt.rydberg

Subscribers: bengt.rydberg, jenkins

Tags: #odin

Maniphest Tasks: T781

Differential Revision: https://phabricator.molflow.com/D685