Summary: CO mode frequency correction of each spectrum in scan
Summary:
modification of the frequency correction of the CO mode. Now the correction is done in two steps.
First an estimate of a scalar frequency offset based on averaged scan data is done(as before).
Then a frequency offset for each spectrum in scan is estimated based on a routine described by Fransesco.

Test Plan: run_all_test.sh

Reviewers: skymandr, joakim

Reviewed By: skymandr

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D818