Repaired error propagation for L2ViewDay
Summary: The Level2ViewDay endpoint didn't propagate errors properly, which mean that the "Too broad query error" caused a 500 instead of a 400. It may be, that the "Too broad query error" should be relaxed and defaults set in some cases, but that is not done in this diff so far.

Test Plan: jenkins

Reviewers: #odin, joakim

Reviewed By: #odin, joakim

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D924