Update Odin API requirements
Summary:
Only do minimal upgrade to the production requirements to avoid potential issues: flask is only upgraded 0.12.3 which should fix the validation vulnerability in the installed version

Rebase `test-requirements.in` on `odinapi-requirement.txt`. This ensure that we test the code with the same library versions used in production.

Test Plan: Run existing tests

Reviewers: #odin, martin

Reviewed By: #odin, martin

Subscribers: jenkins

Differential Revision: https://phabricator.molflow.com/D1087