Integrating with api
Summary:
This includes both reader changes and the api integration (because enough is enough).

Some heuristics in the code for reading the Bdx-data isn't valid for the MIPAS (i.e. one has pressure the other alititude). So I made it work for both.
That means I also extended the API response to include 'Altitude'.

Test Plan:
Adding unit-tests.

The read-function first was covered with the new tests before refactoring.

All apriori mat-files in qsmr-data repo have been manually tested to work with the new reading.

Reviewers: #odin, bengt.rydberg

Reviewed By: #odin, bengt.rydberg

Subscribers: bengt.rydberg, jenkins

Differential Revision: https://phabricator.molflow.com/D1214