a first step towards building odin-api ui with npm and webpack.
important consideration is also that we update jquery to
3.5.0 since there was a security issue with previous versions.

so far this diff does not cover an automatic build.

but I have included the bundle js file
(src/odinapi/static/js/odinlib.js)
created with npm and webpack,
and this is currently loaded in base.html
so the change should be applied in system tests.

It seems that the webpages works as before.