a first step towards building odin-api using webpack. (#35)
* a first step towards building odin-api ui with npm and webpack.

important consideration is also that we update jquery to
3.5.0 since there was a security issue with previous versions.

so far this diff does not cover an automatic build.

but I have included the bundle js file
(src/odinapi/static/js/odinlib.js)
created with npm and webpack,
and this is currently loaded in base.html
so the change should be applied in system tests.

It seems that the webpages works as before.

* fixing failing tests

* fixing problems and removing deps

* now running karma tests on the actual .js file used and building
bundle automatically. a remaining problem is that the
current karma setup does not allow that we test
files where use require and export, which are node
stuff that we can not do in the browser. to allow
this we must modify karma setup, i have not get that
to work.

* modify jenkinsfile so we now first build the bundle
with webpack and the run unit tests on this bundle.
In this way I can re-add the tests that I had
previously removed, since these can now be tested.
Maybe not the most beautiful solution...

* relacing require with import

* removing stuff not needded anymore